Refactored noSession
