+++ lib/deploy.js
@@ -15,52 +15,7 @@ var register = require('./user/create');
 var auto = require('./util/autogenerate');
 var link = require('./project/link');
 var npm = require('./util/npm');
-
-
-/**
- * noSession
- *
- * Attempts to login user or create ghost user for Hub Services when no session is detected.
- */
-var noSession = function(callback) {
-    var loginPrompt = {
-        type: 'confirm',
-        name: 'login',
-        message: 'No current user session. Login to an existing Famous Cloud Services account?: '
-    };
-
-    async.waterfall([
-        function(cb){
-            inquirer.prompt([loginPrompt], function(answers) {
-                cb(null, answers);
-            });
-        },
-        function(response, cb){
-            if (response.login) {
-                loginFlow(function(error, data) {
-                    cb(error, data);
-                });
-            } else {
-                auto.generate(function(error, data){
-                    createUser(data, function(err) {
-                        cb(err, {email: data.email, password: data.password});
-                    });
-                });
-            }
-        },
-        login,
-        storage.setGlobal,
-        function (data, cb) {
-            storage.getGlobal(cb);
-        }
-    ],
-    function (error, data) {
-        if (error === 'login') {
-            callback(null, data);
-        }
-        callback(error, data);
-    });
-};
+var noSession = require('./user/noSession');
 
 /**
  * deployProject
@@ -68,7 +23,6 @@ var noSession = function(callback) {
  * Uploads project to Hub Services.
  */
 var deployProject = function() {
-    console.log('hellor')
     async.waterfall([
         function(callback) {
             fs.readdir(process.cwd(), function (err, files) {
@@ -140,4 +94,3 @@ var deployProject = function() {
 /** **/
 
 module.exports = deployProject;
-
