+++ lib/user/create.js
@@ -15,27 +15,27 @@ var chalk = require('chalk');
  */
 var createUserCLI = function() {
     var emailPrompt = {
-type: 'input',
-      name: 'email',
-      message: 'Email: '
+        type: 'input',
+        name: 'email',
+        message: 'Email: '
     };
 
     var usernamePrompt = {
-type: 'input',
-      name: 'username',
-      message: 'User Name: '
+        type: 'input',
+        name: 'username',
+        message: 'User Name: '
     };
 
     var passPrompt = {
-type: 'password',
-      name: 'password',
-      message: 'Password: '
+        type: 'password',
+        name: 'password',
+        message: 'Password: '
     };
 
     var eulaPrompt = {
-type: 'confirm',
-      name: 'eula',
-      message: "View the Famous Cloud Services TOS at 'http://famous.org/terms.html', do you accept"
+        type: 'confirm',
+        name: 'eula',
+        message: "View the Famous Cloud Services TOS at 'http://famous.org/terms.html', do you accept"
     };
 
     async.waterfall([
